pyproject.toml
[project]
name = "pydanticaiconcept"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai[logfire]>=0.5.0",
    "pydantic-ai-slim[openai]>=0.5.0",
    "pymysql>=1.1.1",
]
